requirements.txt
rembg==2.0.50
Pillow==9.3.0
streamlit==1.21.0
numba==0.57
numpy==1.24.4